code_generator_funcadl_xAOD/pyproject.toml
[tool.poetry]
name = "code-generator-funcadl-xaod"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "code_generator_funcadl_xaod"}]

[tool.poetry.dependencies]
python = "~3.10"
func-adl-xAOD = "2.0.1"
servicex-code-gen-lib = {version = "^1.1.4a10"}

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest-flask = "^1.2.0"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
pytest = "^7.1.3"
autopep8 = "^1.7.0"
flake8 = "^5.0.4"
coverage = "^6.5.0"
codecov = "^2.1.12"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
